test: add benchmark for log processing
In production, we're seeing memory usage increasing heavily when Actions
workflow volumes increases. Here we add a benchmark to reproduce that,
so we can work on fixes.

It's a table benchmark which runs for different sizes of logs. We feed
these logs to handler and benchmark the processing of them.