chore: upgrade Go to 1.24.5 and golangci-lint v2 (#343)
Update all go.mod files, Makefile.Common, and golangci.yaml to use Go
1.24.5. This enables the new testing.B.Loop() functionality for more
robust benchmarking.

We needed to bump golangci-lint to be compatible with Go 1.24.x. The
latest version is a major version bump with new package names. It found
some lints to fix, which are included in this commit.